+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@fortawesome/fontawesome-free": "^6.2.0",
     "bootstrap": "^5.2.1",
+    "dotenv": "^16.0.3",
     "firebase": "^9.9.4",
     "react": "^18.2.0",
     "react-bootstrap": "^2.5.0",
