+++ src/components/Router.tsx
@@ -1,4 +1,3 @@
-import React, { useState } from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { Auth } from "../routes/Auth";
 import { Home } from "../routes/Home";
