+++ src/routes/Auth.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import { firebaseInstance } from "../firebase";
 
 export const Auth = () => {
   const onSocialClick = (event: any) => {};
